+++ cognite/neat/_session/_base.py
@@ -13,6 +13,7 @@ from cognite.neat._rules.transformers import ConvertToRules, VerifyAnyRules
 
 from ._prepare import PrepareAPI
 from ._read import ReadAPI
+from ._set import SetAPI
 from ._show import ShowAPI
 from ._state import SessionState
 from ._to import ToAPI
@@ -34,6 +35,7 @@ class NeatSession:
         self.to = ToAPI(self._state, client, verbose)
         self.prepare = PrepareAPI(self._state, verbose)
         self.show = ShowAPI(self._state)
+        self.set = SetAPI(self._state, verbose)
 
     @property
     def version(self) -> str: